requirements.txt
flask==2.0.2
Flask-SQLAlchemy==2.5.1
python-decouple==3.5
gunicorn==20.1.0
Flask-Migrate==3.1.0
psycopg2-binary==2.9.1
marshmallow==3.13.0
flask-admin==1.5.8
pytest==6.2.5
pytest-mock==3.6.1
isort==5.9.3
